Add "development" tests
