+++ packages/react-scripts/templates/kitchensink/integration/README.md
@@ -1,0 +1,4 @@
+# IMPORTANT
+
+Until further investigations, there is a max limit of `initDOM` in one single file.
+Empirically they should not exceed 10 tests per file.
